feat: add support for ping events in Twitch Chat
